user_service/requirements.txt
Flask==1.0.2
flask-restful==0.3.7
flask-marshmallow==0.9.0
mysql-connector
flask-cors